Reduced memory usage, simplified topic structure

+++ coordinator/src/zigbee/XbeeSerialHandler.cpp
@@ -16,8 +16,8 @@ bool XbeeSerialHandler::isFrameAvailable(){
 
 XbeeReading XbeeSerialHandler::createReadingFromSerial(){
   // byte receivedMac[8];
-  int analogMSB;
-  int analogLSB;
+  byte analogMSB;
+  byte analogLSB;
 
   // if(Serial.available() >= ZIGBEE_FRAME_SIZE){
   //   if(Serial.read() != ZIGBEE_START_BYTE) {
